Fix issue with the sizing of the text input field when the visibility of the input control is dynamically toggled, for example via the `ng-disabled` attribute on the `<ui-select>` element.
The problem is caused by the width calculation getting an incorrect 0 `offsetLeft` value for the input, as the `offsetParent` of the input is still `null` when the controller tries to calculate the width. The fix resolves this by adding a watch for the `offsetParent`, and only performing the width calculation when both the `offsetParent` *and* the container width are available.